+++ examples/example_3_tcp.cpp
@@ -1,12 +1,18 @@
 /**
  * Example of the SIM5320E usage with STM32F3Discovery board.
  *
- * The example shows TCP usage.
+ * The example of the TCP usage. It uses HTTP to download and show web page.
+ *
+ * Requirements:
+ *
+ * - active SIM card with an internet access
  *
  * Pin map:
  *
  * - PB_10 - UART TX (SIM5320E)
  * - PB_11 - UART RX (SIM5320E)
+ *
+ * Note: to run the example, you should an adjust APN settings in the code.
  */
 
 #include "mbed.h"
@@ -18,30 +24,17 @@
 using namespace sim5320;
 
 #define APP_ERROR(err, message) MBED_ERROR(MBED_MAKE_ERROR(MBED_MODULE_APPLICATION, err), message)
-#define CHECK_RET_CODE(expr)                                                                              \
-    {                                                                                                     \
-        int err = expr;                                                                                   \
-        if (err < 0) {                                                                                    \
-            MBED_ERROR(MBED_MAKE_ERROR(MBED_MODULE_APPLICATION, err), "Expression \"" #expr "\" failed"); \
-        }                                                                                                 \
+#define CHECK_RET_CODE(expr)                                                           \
+    {                                                                                  \
+        int err = expr;                                                                \
+        if (err < 0) {                                                                 \
+            char err_msg[64];                                                          \
+            sprintf(err_msg, "Expression \"" #expr "\" failed (error code: %i)", err); \
+            MBED_ERROR(MBED_MAKE_ERROR(MBED_MODULE_APPLICATION, err), err_msg);        \
+        }                                                                              \
     }
 
 DigitalOut led(LED2);
-
-static const char *get_reg_mode_name(CellularNetwork::RegistrationType type)
-{
-
-    switch (type) {
-    case CellularNetwork::C_EREG:
-        return "C_EREG";
-    case CellularNetwork::C_GREG:
-        return "C_GREG";
-    case CellularNetwork::C_REG:
-        return "C_REG";
-    default:
-        return "Unknown error";
-    }
-}
 
 static const char *get_reg_status_name(CellularNetwork::RegistrationStatus status)
 {
@@ -192,7 +185,6 @@ void print_http_page(NetworkInterface *network, const char *host, int port, cons
     CHECK_RET_CODE(socket.close());
 }
 
-// simple led demo
 int main()
 {
     // create driver
@@ -209,6 +201,7 @@ int main()
 
     // set credential
     //context->set_sim_pin("1234");
+    // note: set your APN parameters
     context->set_credentials("internet.mts.ru", "mts", "mts");
     // connect to network
     CHECK_RET_CODE(context->connect()); // note: by default operations is blocking