+++ include/sim5320_utils.h
@@ -6,6 +6,8 @@
 #include "stdarg.h"
 namespace sim5320 {
 
+static const int SIM5320_DEFAULT_TIMEOUT = 8000;
+
 /**
  * Helper function that to return first found error code.
  *
@@ -23,11 +25,6 @@ inline nsapi_error_t any_error(nsapi_error_t err_1, nsapi_error_t err_2)
     }
     return 0;
 }
-
-#define SIM5320_UNLOCK_RETURN_IF_ERROR(at_handler_ptr) \
-    if (at_handler_ptr->get_last_error()) {            \
-        return at_handler_ptr->unlock_return_error();  \
-    }
 
 /**
  * Read AT response that: